* changed to use C99 variable types (int8_t, uint8_t, etc) 	* added CButtonState class which replaces the buttons parameter in CView mouse methods 	* Auto drag scrolling flag for CScrollView 	* removed CGraphicsPath::addString method 	* added some new draw styles to CKnob : 		- drawing a circle instead of lines for the handle 		- drawing the value as a corona (with support for inverted value and value from center) 	* added drawing style to CSlider 	* simplified stringToValue and valueToString functions for CParamDisplay and CTextEdit

+++ vstgui/doxygen/page_tutorial.h
@@ -173,7 +173,7 @@ Now in our valueChanged method we can differentiate between the controls :
 @code
 void MyEditor::valueChanged (CControl* pControl)
 {
-	long tag = pControl->getTag ();
+	int32_t tag = pControl->getTag ();
 	if (tag == 0)
 	{
 		// the first knob
